update DownloadFile class.

+++ EnumRun/ScriptDelivery/DeleteManager.cs
@@ -67,6 +67,8 @@ namespace EnumRun.ScriptDelivery
 
         private void SearchTarget()
         {
+            if (!Directory.Exists(_targetDir)) { return; }
+
             List<SearchPath> _targetList = new List<SearchPath>(Targetlist.Select(x => new SearchPath(_targetDir, x)));
             List<SearchPath> _excludeList = new List<SearchPath>(ExcludeList.Select(x => new SearchPath(_targetDir, x)));
 
@@ -120,31 +122,16 @@ namespace EnumRun.ScriptDelivery
 
         private void DeleteTarget()
         {
-            foreach (string delTarget in _dList)
+            foreach (string delTarget in _dList ?? new List<string>())
             {
                 try
                 {
-                    string destination = Path.Combine(_trashPath, Path.GetRelativePath(_targetDir, delTarget));
-                    string parent = Path.GetDirectoryName(destination);
-                    if (!Directory.Exists(parent))
+                    if (_trashPath == null)
                     {
-                        Directory.CreateDirectory(parent);
+                        Directory.Delete(delTarget, true);
+                        Console.WriteLine("▲;[Del] {0}", delTarget);
                     }
-                    if (Directory.Exists(destination))
-                    {
-                        Directory.Delete(destination, true);
-                    }
-                    Directory.Move(delTarget, destination);
-
-                    Console.WriteLine("▲;[Del] {0}", destination);
-                }
-                catch { }
-            }
-            foreach (string delTarget in _fList)
-            {
-                if (File.Exists(delTarget))
-                {
-                    try
+                    else
                     {
                         string destination = Path.Combine(_trashPath, Path.GetRelativePath(_targetDir, delTarget));
                         string parent = Path.GetDirectoryName(destination);
@@ -152,13 +139,44 @@ namespace EnumRun.ScriptDelivery
                         {
                             Directory.CreateDirectory(parent);
                         }
-                        if (File.Exists(destination))
+                        if (Directory.Exists(destination))
                         {
-                            File.Delete(destination);
+                            Directory.Delete(destination, true);
                         }
-                        File.Move(delTarget, destination);
+                        Directory.Move(delTarget, destination);
 
-                        Console.WriteLine("▲:[Del] {0}", destination);
+                        Console.WriteLine("▲;[Del] {0} -> {1}", delTarget, destination);
+                    }
+                }
+                catch { }
+            }
+            foreach (string delTarget in _fList ?? new List<string>())
+            {
+                if (File.Exists(delTarget))
+                {
+                    try
+                    {
+                        if (_trashPath == null)
+                        {
+                            File.Delete(delTarget);
+                            Console.WriteLine("▲;[Del] {0}", delTarget);
+                        }
+                        else
+                        {
+                            string destination = Path.Combine(_trashPath, Path.GetRelativePath(_targetDir, delTarget));
+                            string parent = Path.GetDirectoryName(destination);
+                            if (!Directory.Exists(parent))
+                            {
+                                Directory.CreateDirectory(parent);
+                            }
+                            if (File.Exists(destination))
+                            {
+                                File.Delete(destination);
+                            }
+                            File.Move(delTarget, destination);
+
+                            Console.WriteLine("▲:[Del] {0} -> {1}", delTarget, destination);
+                        }
                     }
                     catch { }
                 }
